+++ template/f-debug/src/main/java/ru/surfstudio/standard/f_debug/DebugInteractor.kt
@@ -1,6 +1,7 @@
 package ru.surfstudio.standard.f_debug
 
 import android.app.Application
+import com.codemonkeylabs.fpslibrary.TinyDancer
 import com.readystatesoftware.chuck.ChuckInterceptor
 import com.squareup.leakcanary.LeakCanary
 import okhttp3.OkHttpClient
@@ -9,6 +10,7 @@ import ru.surfstudio.android.dagger.scope.PerApplication
 import ru.surfstudio.standard.f_debug.notification.DebugNotificationBuilder
 import ru.surfstudio.standard.f_debug.server_settings.reboot.interactor.RebootInteractor
 import ru.surfstudio.standard.f_debug.storage.DebugServerSettingsStorage
+import ru.surfstudio.standard.f_debug.storage.DebugUiToolsStorage
 import ru.surfstudio.standard.f_debug.storage.MemoryDebugStorage
 import javax.inject.Inject
 
@@ -16,6 +18,7 @@ import javax.inject.Inject
 class DebugInteractor @Inject constructor(
         private val memoryDebugStorage: MemoryDebugStorage,
         private val debugServerSettingsStorage: DebugServerSettingsStorage,
+        private val debugUiToolsStorage: DebugUiToolsStorage,
         private val application: Application,
         private val rebootInteractor: RebootInteractor
 ) {
@@ -25,6 +28,12 @@ class DebugInteractor @Inject constructor(
         get() = memoryDebugStorage.isLeakCanaryEnabled
         set(value) {
             memoryDebugStorage.isLeakCanaryEnabled = value
+        }
+
+    var isFpsEnabled: Boolean
+        get() = debugUiToolsStorage.isFpsEnabled
+        set(value) {
+            debugUiToolsStorage.isFpsEnabled = value
         }
 
     /**
@@ -65,6 +74,10 @@ class DebugInteractor @Inject constructor(
     fun configureOkHttp(okHttpBuilder: OkHttpClient.Builder) {
         if (debugServerSettingsStorage.isChuckEnabled) {
             okHttpBuilder.addInterceptor(ChuckInterceptor(application))
+        }
+
+        if (debugUiToolsStorage.isFpsEnabled) {
+            TinyDancer.create().show(application)
         }
     }
     //endregion