+++ template/f-debug/src/main/java/ru/surfstudio/standard/f_debug/debug/DebugActivityView.kt
@@ -45,5 +45,7 @@ class DebugActivityView : BaseRenderableActivityView<DebugScreenModel>() {
         show_fcm_token_item_layout.setOnClickListener { presenter.openFcmTokenScreen() }
         show_memory_item_layout.setOnClickListener { presenter.openMemoryScreen() }
         show_app_info_item_layout.setOnClickListener { presenter.openAppInfoScreen() }
+        show_ui_tools_item_layout.setOnClickListener { presenter.openUiToolsScreen() }
+        show_developer_tools_item_layout.setOnClickListener { presenter.openDeveloperToolsScreen() }
     }
 }